lang/eval: ModuleInstance exprs.Scope methods in separate file
The module_instance.go file was getting quite long and hard to navigate,
and the Valuer vs. Scope distinction seems like a reasonable place to
draw a line between two different files related to module instances.
